Estimate or Actual File Upload Processing code removal and names refactoring
