+++ ChargesApi.Tests/V1/Controllers/EstimatesUploadControllerTests.cs
@@ -19,21 +19,21 @@ namespace ChargesApi.Tests.V1.Controllers
     [Collection("LogCall collection")]
     public class EstimatesUploadControllerTests
     {
-        private readonly EstimatesUploadController _estimatesUploadController;
+        private readonly EstimatesActualUploadController _estimatesUploadController;
         private readonly ControllerContext _controllerContext;
         private readonly HttpContext _httpContext;
         private readonly Fixture _fixture;
 
-        private readonly Mock<IAddEstimateChargesUseCase> _mockAddEstimateChargesUseCase;
+        private readonly Mock<IEstimateActualUploadUseCase> _mockAddEstimateChargesUseCase;
         //private const string Token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJ0ZXN0IiwiaWF0IjoxNjM5NDIyNzE4LCJleHAiOjE5ODY1Nzc5MTgsImF1ZCI6InRlc3QiLCJzdWIiOiJ0ZXN0IiwiZ3JvdXBzIjpbInNvbWUtdmFsaWQtZ29vZ2xlLWdyb3VwIiwic29tZS1vdGhlci12YWxpZC1nb29nbGUtZ3JvdXAiXSwibmFtZSI6InRlc3RpbmcifQ.IcpQ00PGVgksXkR_HFqWOakgbQ_PwW9dTVQu4w77tmU";
 
         public EstimatesUploadControllerTests()
         {
             _fixture = new Fixture();
-            _mockAddEstimateChargesUseCase = new Mock<IAddEstimateChargesUseCase>();
+            _mockAddEstimateChargesUseCase = new Mock<IEstimateActualUploadUseCase>();
             _httpContext = new DefaultHttpContext();
             _controllerContext = new ControllerContext(new ActionContext(_httpContext, new RouteData(), new ControllerActionDescriptor()));
-            _estimatesUploadController = new EstimatesUploadController(_mockAddEstimateChargesUseCase.Object)
+            _estimatesUploadController = new EstimatesActualUploadController(_mockAddEstimateChargesUseCase.Object)
             {
                 ControllerContext = _controllerContext
             };
