+++ ChargesApi.Tests/V1/UseCase/EstimateActualUploadUseCaseTests.cs
@@ -11,28 +11,24 @@ using ChargesApi.V1.Factories;
 
 namespace ChargesApi.Tests.V1.UseCase
 {
-    public class AddEstimateChargesUseCaseTests
+    public class EstimateActualUploadUseCaseTests
     {
-        private readonly Mock<IChargesApiGateway> _mockChargeGateway;
         private readonly Mock<IHousingSearchService> _mockHousingSearchService;
-        private readonly Mock<IFinancialSummaryService> _mockFinancialService;
         private readonly Mock<IAwsS3FileService> _mockS3FileService;
         private readonly Mock<ISnsGateway> _mockSnsGateway;
         private readonly Mock<ISnsFactory> _mockSnsFactory;
-        private readonly Mock<ILogger<AddEstimateChargesUseCase>> _mockLogger;
+        private readonly Mock<ILogger<EstimateActualUploadUseCase>> _mockLogger;
 
         //private const string Token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJ0ZXN0IiwiaWF0IjoxNjM5NDIyNzE4LCJleHAiOjE5ODY1Nzc5MTgsImF1ZCI6InRlc3QiLCJzdWIiOiJ0ZXN0IiwiZ3JvdXBzIjpbInNvbWUtdmFsaWQtZ29vZ2xlLWdyb3VwIiwic29tZS1vdGhlci12YWxpZC1nb29nbGUtZ3JvdXAiXSwibmFtZSI6InRlc3RpbmcifQ.IcpQ00PGVgksXkR_HFqWOakgbQ_PwW9dTVQu4w77tmU";
         private readonly Fixture _fixture;
 
-        private readonly AddEstimateChargesUseCase _addEstimateChargesUseCase;
+        private readonly EstimateActualUploadUseCase _addEstimateChargesUseCase;
 
-        public AddEstimateChargesUseCaseTests()
+        public EstimateActualUploadUseCaseTests()
         {
             _fixture = new Fixture();
-            _mockChargeGateway = new Mock<IChargesApiGateway>();
             _mockHousingSearchService = new Mock<IHousingSearchService>();
-            _mockLogger = new Mock<ILogger<AddEstimateChargesUseCase>>();
-            _mockFinancialService = new Mock<IFinancialSummaryService>();
+            _mockLogger = new Mock<ILogger<EstimateActualUploadUseCase>>();
             _mockS3FileService = new Mock<IAwsS3FileService>();
             _mockSnsGateway = new Mock<ISnsGateway>();
             _mockSnsFactory = new Mock<ISnsFactory>();
@@ -44,9 +40,7 @@ namespace ChargesApi.Tests.V1.UseCase
                     BucketName = "test-bucket",
                     FileUrl = null
                 });
-            _addEstimateChargesUseCase = new AddEstimateChargesUseCase(_mockChargeGateway.Object,
-                _mockHousingSearchService.Object,
-                _mockFinancialService.Object,
+            _addEstimateChargesUseCase = new EstimateActualUploadUseCase(
                _mockLogger.Object,
                 _mockS3FileService.Object,
                 _mockSnsGateway.Object,
