+++ ChargesApi/V1/Boundary/Request/AddEstimatesActualRequest.cs
@@ -5,12 +5,12 @@ using System.ComponentModel.DataAnnotations;
 
 namespace ChargesApi.V1.Boundary.Request
 {
-    public class AddEstimatesRequest
+    public class AddEstimatesActualRequest
     {
         [Required(ErrorMessage = "Please select a file.")]
         [DataType(DataType.Upload)]
         [AllowedExtensions(new string[] { ".xlsx", ".xls" })]
-        public IFormFile EstimatesFile { get; set; }
+        public IFormFile EstimatesActualFile { get; set; }
 
         public ChargeGroup ChargeGroup { get; set; }
     }
