--- ChargesApi/V1/Gateways/Services/Interfaces/IFinancialSummaryService.cs
@@ -1,10 +1,0 @@
-using ChargesApi.V1.Domain;
-using System.Threading.Tasks;
-
-namespace ChargesApi.V1.Gateways.Services.Interfaces
-{
-    public interface IFinancialSummaryService
-    {
-        Task<bool> AddEstimateSummary(AddAssetSummaryRequest addAssetSummaryRequest);
-    }
-}
