+++ ChargesApi/appsettings.json
@@ -15,9 +15,6 @@
     "LocalAccessKey": "2cl9i",
     "LocalSecretKey": "vymxp"
   },
-  "BatchProcessing": {
-    "PerBatchCount": 5000
-  },
   "AllowedHosts": "*",
   "XRay": {
     "DisableXRayTracing": "true",
